Change out React state handling in favor of routing
In order to allow for a slightly more persistent app as well as
reduce the amount of state that needs to be handled at the app-level,
React state handlie (i.e. useState calls) were removed and instead
any state (search query, sorting, etc.) was moved to the browser
level by way of routing.